preguteca_backend/requirements.txt
asgiref==3.6.0
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.1.7
django-browser-reload==1.8.0
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
iso8601==2.0.0
Markdown==3.4.3
metomi-isodatetime==1!3.0.0
mypy==1.2.0
mypy-extensions==1.0.0
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.5.3
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.15
